+++ index.html
@@ -37,7 +37,7 @@
     </div>
 
     <div>
-        Samle Response:
+        Sample Response:
         <textarea class="form-control" aria-label="With textarea"></textarea>
     </div>
 </div>
