--- lib/Member.js
@@ -1,29 +1,0 @@
-'use strict';
-
-export default class Member {
-    constructor(firstName, lastName) {
-        this.firstName = firstName;
-        this.lastName = lastName;
-    }
-
-    getName() {
-        return this.lastName + ' ' + this.firstName;
-    }
-
-    get firstName() {
-        return this._firstName;
-    }
-
-    set firstName(value) {
-        this._firstName = value;
-    }
-
-    get lastName() {
-        return this._lastName;
-    }
-
-    set lastName(value) {
-        this._lastName = value;
-    }
-}
-
